refactor : 쿠폰 발급 전략 패턴
