+++ src/main/java/com/sparta/couponpop/domain/coupon/service/coupon_issue/SynchronizedCouponIssueService.java
@@ -1,0 +1,24 @@
+package com.sparta.couponpop.domain.coupon.service.coupon_issue;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class SynchronizedCouponIssueService implements CouponIssueFacade{
+
+    private final DefaultCouponIssueService defaultCouponIssueService;
+
+    @Override
+    public void issueCoupon(Long memberId, Long eventId, LocalDateTime currentDateTime) {
+        log.info("락 획득");
+        synchronized (this) {
+            defaultCouponIssueService.issueCoupon(memberId, eventId, currentDateTime);
+        }
+        log.info("락 반납. 쿠폰 발급 완료");
+    }
+}
