+++ src/test/java/com/sparta/couponpop/domain/coupon/controller/CouponControllerTest.java
@@ -14,9 +14,8 @@ import com.sparta.couponpop.domain.coupon.dto.response.IssuedCouponListResponse;
 import com.sparta.couponpop.domain.coupon.enums.CouponStatus;
 import com.sparta.couponpop.domain.coupon.exception.CouponErrorCode;
 import com.sparta.couponpop.domain.coupon.repository.dto.CouponSummaryInfoProjection;
-import com.sparta.couponpop.domain.coupon.service.coupon_issue.CouponIssueFacade;
-import com.sparta.couponpop.domain.coupon.service.coupon_issue.CouponIssueService;
 import com.sparta.couponpop.domain.coupon.service.CouponService;
+import com.sparta.couponpop.domain.coupon.service.coupon_issue.CouponIssueFacade;
 import com.sparta.couponpop.domain.couponevent.exception.CouponEventErrorCode;
 import com.sparta.couponpop.domain.member.enums.MemberType;
 import org.junit.jupiter.api.BeforeEach;
